requirements.txt
newsapi-python
pymongo
python-dotenv
